Refactor: split mixin classes into different python modules
The number of mixin classes is growing, split them into different modules
to make them more managable

Refs: LIIK-85
